Move model classes
