+++ src/main/groovy/org/hidetake/groovy/ssh/interaction/InteractionHandler.groovy
@@ -3,7 +3,7 @@ package org.hidetake.groovy.ssh.interaction
 import static org.hidetake.groovy.ssh.util.ClosureUtil.callWithDelegate
 
 /**
- * A handler of the {@link org.hidetake.groovy.ssh.api.OperationSettings#interaction} closure.
+ * A handler of the {@link org.hidetake.groovy.ssh.core.settings.OperationSettings#interaction} closure.
  *
  * @author hidetake.org
  */
