+++ src/main/groovy/org/hidetake/groovy/ssh/session/Plan.groovy
@@ -1,7 +1,7 @@
 package org.hidetake.groovy.ssh.session
 
 import groovy.transform.Canonical
-import org.hidetake.groovy.ssh.api.Remote
+import org.hidetake.groovy.ssh.core.Remote
 
 @Canonical
 class Plan<T> {
