+++ src/main/groovy/org/hidetake/groovy/ssh/session/SessionExecutor.groovy
@@ -1,8 +1,8 @@
 package org.hidetake.groovy.ssh.session
 
 import groovy.util.logging.Slf4j
-import org.hidetake.groovy.ssh.api.CompositeSettings
-import org.hidetake.groovy.ssh.api.ConnectionSettings
+import org.hidetake.groovy.ssh.core.settings.CompositeSettings
+import org.hidetake.groovy.ssh.core.settings.ConnectionSettings
 import org.hidetake.groovy.ssh.connection.ConnectionManager
 import org.hidetake.groovy.ssh.operation.DefaultOperations
 import org.hidetake.groovy.ssh.operation.DryRunOperations
