+++ src/test/groovy/org/hidetake/groovy/ssh/connection/ProxyValidatorSpec.groovy
@@ -1,10 +1,10 @@
 package org.hidetake.groovy.ssh.connection
 
-import org.hidetake.groovy.ssh.api.Proxy
+import org.hidetake.groovy.ssh.core.Proxy
 import spock.lang.Specification
 
-import static org.hidetake.groovy.ssh.api.ProxyType.HTTP
-import static org.hidetake.groovy.ssh.api.ProxyType.SOCKS
+import static org.hidetake.groovy.ssh.core.ProxyType.HTTP
+import static org.hidetake.groovy.ssh.core.ProxyType.SOCKS
 
 class ProxyValidatorSpec extends Specification {
