+++ src/test/groovy/org/hidetake/groovy/ssh/core/ServiceSpec.groovy
@@ -1,7 +1,6 @@
 package org.hidetake.groovy.ssh.core
 
-import org.hidetake.groovy.ssh.api.OperationSettings
-import org.hidetake.groovy.ssh.api.Remote
+import org.hidetake.groovy.ssh.core.settings.OperationSettings
 import org.hidetake.groovy.ssh.session.Plan
 import org.hidetake.groovy.ssh.session.SessionExecutor
 import spock.lang.Specification
