+++ src/test/groovy/org/hidetake/groovy/ssh/server/CommandExecutionSpec.groovy
@@ -5,7 +5,7 @@ import org.apache.sshd.server.CommandFactory
 import org.apache.sshd.server.PasswordAuthenticator
 import org.codehaus.groovy.tools.Utilities
 import org.hidetake.groovy.ssh.Ssh
-import org.hidetake.groovy.ssh.api.OperationSettings
+import org.hidetake.groovy.ssh.core.settings.OperationSettings
 import org.hidetake.groovy.ssh.core.Service
 import org.hidetake.groovy.ssh.session.BadExitStatusException
 import org.hidetake.groovy.ssh.operation.DefaultOperations
