+++ src/test/groovy/org/hidetake/groovy/ssh/server/ShellExecutionSpec.groovy
@@ -4,7 +4,7 @@ import org.apache.sshd.SshServer
 import org.apache.sshd.common.Factory
 import org.apache.sshd.server.PasswordAuthenticator
 import org.hidetake.groovy.ssh.Ssh
-import org.hidetake.groovy.ssh.api.OperationSettings
+import org.hidetake.groovy.ssh.core.settings.OperationSettings
 import org.hidetake.groovy.ssh.core.Service
 import org.hidetake.groovy.ssh.session.BadExitStatusException
 import org.hidetake.groovy.ssh.operation.DefaultOperations
